refactor: ViewController 내 타이머 관련 기능 분리 - TimerStackView 타입 구현
